Timezone offset needs a cast.

+++ event.go
@@ -30,7 +30,7 @@ type Event struct {
 func getWeekFromTimestamp(ts time.Time) int64 {
 	var offset int
 	_, offset = ts.Zone()
-	return (ts.Unix() + offset) / (7 * 24 * 60 * 60)
+	return (ts.Unix() + int64(offset)) / (7 * 24 * 60 * 60)
 }
 
 // Create a new event with the speicfied details.
